requirements.txt
requests==2.31.0
python-dateutil==2.8.2
PyGithub==2.1.1
boto3>=1.34.0 
